R-tree short -> int wip

+++ sea.hpp
@@ -20,7 +20,7 @@ namespace ss {
         void teleport_to(int id, float x, float y, float vx = 0, float vy = 0);
         void teleport_to(const char* guid, float x, float y, float vx = 0, float vy = 0);
         void teleport_by(const char* guid, float dx, float dy);
-        void query_near_lng_lat_to_packet(float lng, float lat, short halfex, std::vector<sea_object_public>& sop_list) const;
+        void query_near_lng_lat_to_packet(float lng, float lat, int halfex, std::vector<sea_object_public>& sop_list) const;
         void query_near_to_packet(float xc, float yc, float ex, std::vector<sea_object_public>& sop_list) const;
         void update(float delta_time);
         void set_object_state(int id, SEA_OBJECT_STATE state);
@@ -35,8 +35,8 @@ namespace ss {
         std::unordered_map<int, sea_object> sea_objects;
         std::unordered_map<std::string, int> sea_guid_to_id;
         bgi::rtree< value, bgi::quadratic<16> > rtree;
-        const short res_width;
-        const short res_height;
+        const int res_width;
+        const int res_height;
         const float km_per_cell;
         int tick_seq;
     };
